// frontend/src/app/core/api/model/createShareRequest.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateShareRequest { 
    vaultEntryId: number;
    recipientEmail?: string;
    expiryHours?: number;
    maxViews?: number;
    permission?: string;
}